% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{en.vir}
\alias{en.vir}
\title{en.vir data set, a class of RasterBrick.}
\format{
a class of RasterBrick.
\describe{
\item{en.vir}{class: RasterBrick;
dimensions : 6, 2160, 12960, 5  (nrow, ncol, ncell, nlayers);
resolution : 0.1666667, 0.1666667  (x, y);
extent: -180, 180, -60, 90  (xmin, xmax, ymin, ymax);
crs:+proj=longlat +datum=WGS84;
source:memory;
names: layer.1, layer.2, ..., .
}
}
}
\source{
\url{http://www.worldclim.org/}
}
\usage{
en.vir
}
\description{
A dataset containing 5 of the 19 bioclimatic variables downloaded from WorldClim
(version 1.4 with 2.5 arc minute resolution; Hijmans et al. 2005)).
}
\keyword{datasets}
